python-code/requirements-docker.txt
appdirs==1.4.4
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
black==20.8b1
bleach==3.3.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
cloudpickle==1.6.0
coverage==5.4
cycler==0.10.0
dask==2021.4.1
DBUtils==2.0
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
flake8==3.9.1
fsspec==2021.4.0
funcsigs==1.0.2
glob2==0.7
hupper==1.10.2
hurry.filesize==0.9
idna==2.10
iniconfig==1.1.1
ipykernel==5.4.2
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
isort==5.8.0
jedi==0.17.2
Jinja2==2.11.3
jmetalpy==1.5.5
joblib==1.0.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
llvmlite==0.36.0
locket==0.2.1
Mako==1.1.4
MarkupSafe==1.1.1
matplotlib==3.3.3
mccabe==0.6.1
mistune==0.8.4
mypy-extensions==0.4.3
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
notebook==6.2.0
numba==0.53.1
numexpr==2.7.2
numpy==1.19.5
orca==1.5.4
packaging==20.9
pandas==1.2.0
pandocfilters==1.4.3
parso==0.7.1
partd==1.2.0
PasteDeploy==2.1.1
pathspec==0.8.1
patsy==0.5.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
plaster==1.0
plaster-pastedeploy==0.7
plotly==4.14.3
pluggy==0.13.1
prometheus-client==0.9.0
prompt-toolkit==3.0.8
psutil==5.8.0
ptyprocess==0.6.0
py==1.10.0
py4j==0.10.9
pyarrow==3.0.0
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
Pygments==2.7.3
pyparsing==2.4.7
pyramid==2.0
pyrsistent==0.17.3
pyspark==3.0.0
pyspark-stubs==3.0.0.post1
pytest==6.2.2
pytest-cov==2.11.1
python-dateutil==2.8.1
pytools==2021.2.3
pytz==2020.5
PyYAML==5.4.1
pyzmq==20.0.0
qtconsole==5.0.2
QtPy==1.9.0
recmetrics @ git+https://github.com/statisticianinstilettos/recmetrics@e8d9b39131999c484c4c98265f91e233be6ca4cb
regex==2021.4.4
requests==2.25.1
retrying==1.3.3
scikit-learn==0.23.2
scipy==1.6.0
seaborn==0.11.1
Send2Trash==1.5.0
six==1.15.0
statsmodels==0.12.2
surprise==0.1
tables==3.6.1
terminado==0.9.2
testpath==0.4.4
threadpoolctl==2.1.0
toml==0.10.2
toolz==0.11.1
tornado==6.1
tqdm==4.57.0
traitlets==5.0.5
translationstring==1.4
typed-ast==1.4.3
typing-extensions==3.7.4.3
urllib3==1.26.3
venusian==3.0.0
wcwidth==0.2.5
webencodings==0.5.1
WebOb==1.8.7
widgetsnbextension==3.5.1
zope.deprecation==4.4.0
zope.interface==5.4.0
